+++ demos/transitions.cpp
@@ -67,7 +67,7 @@ struct TransitionsDemoScene : Scene {
         }
     }
 
-    void update(uint32_t dt) override
+    void update(double dt) override
     {
         for (auto const& event : this->get_events()) {
             auto keyboard_key = event.keyboard_key();
